fix: correct LM Studio model IDs and status LIMIT overflow
- Default embedding model: text-embedding-nomic-embed-text-v1.5@q8_0
  (LM Studio requires full model identifier)
- Default LLM: qwen/qwen3-30b-a3b for distillation
- Fix rem status crash: usize::MAX overflows i64 in SQL LIMIT

+++ cli/src/main.rs
@@ -557,7 +557,7 @@ fn cmd_status() -> Result<()> {
         // Row counts
         match open_store(&config) {
             Ok(store) => {
-                let sessions = store.get_recent_sessions(usize::MAX)?;
+                let sessions = store.get_recent_sessions(i32::MAX as usize)?;
                 println!("  Sessions:    {}", sessions.len());
 
                 let memories = store.search_memories("")?;
